requirements.txt
aiohttp==3.6.2
aiohttp-jinja2==1.2.0
aiomysql==0.0.20
appdirs==1.4.3
astroid==2.3.3
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
backcall==0.1.0
cffi==1.14.0
chardet==3.0.4
cleo==0.7.6
click==7.1.1
clikit==0.4.2
colorlog==4.1.0
cryptography==2.8
cssselect==1.1.0
decorator==4.4.2
demjson==2.2.4
idna==2.9
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.16.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
lxml==4.5.0
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
multidict==4.7.5
parsel==1.5.2
parso==0.6.2
pastel==0.2.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycparser==2.20
pyee==7.0.1
Pygments==2.6.1
pylev==1.3.0
pylint==2.4.4
PyMySQL==0.9.2
pyppeteer==0.0.25
pytesseract==0.3.3
PyYAML==5.3
selectorlib==0.16.0
six==1.14.0
tqdm==4.43.0
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.8
uvloop==0.14.0
w3lib==1.21.0
wcwidth==0.1.9
websockets==6.0
wrapt==1.11.2
yarl==1.4.2
